Update code and move modeling.py to utils folder. (#290)
* Update code and move modeling.py to utils folder.

* move modeling.py to generation folder

* Fixed typo

* Update code for support instance of fx model to TracedModelForCausalLM

* Update code

* Fixed typo

* Update readme
