--- examples/ipex/text-generation/README.md
@@ -1,30 +1,0 @@
-<!---
-Copyright 2023 The HuggingFace Team. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-
-## Language generation
-
-Based on the script [`run_generation.py`](https://github.com/huggingface/optimum-intel/blob/main/examples/ipex/text-generation/run_generation.py).
-
-Conditional text generation using the auto-regressive models sucha s : GPT, GPT-2, Transformer-XL, XLNet, CTRL. By setting `--jit` an instance of `torch.jit.ScriptModule` will be created and used to perform generation.
-Example usage:
-
-```bash
-python run_generation.py \
-    --model_type gpt2 \
-    --model_name_or_path gpt2 \
-    --output_dir model \
-    --jit
-```
