--- examples/ipex/text-generation/requirements.txt
@@ -1,3 +1,0 @@
-sentencepiece != 0.1.92
-protobuf
-torch >= 2.0.0
