+++ examples/neural_compressor/text-generation/requirements.txt
@@ -1,0 +1,3 @@
+sentencepiece != 0.1.92
+protobuf
+torch >= 2.0.0
